[fix] fix unsigned type
in gyro data not unsigned data.

+++ kalmanfilter_2d/kalman.c
@@ -14,7 +14,7 @@ __weak void kalmanInit(Kalman_t *obj)
     obj->R[1] = 0.7050;
 }
 
-int32_t kalmanfilter(Kalman_t *obj, uint16_t sen1, uint16_t sen2)
+int32_t kalmanfilter(Kalman_t *obj, int32_t sen1, int32_t sen2)
 {
     double var = 0;
     double var1 = (double)sen1;
